nimble/controller: Add CSA #2 to supported features
Feature bit should be set if CSA #2 is supported.

< HCI Command: LE Read Local Supported Features (0x08|0x0003) plen 0
> HCI Event: Command Complete (0x0e) plen 12
      LE Read Local Supported Features (0x08|0x0003) ncmd 1
        Status: Success (0x00)
        Features: 0xfd 0x40 0x00 0x00 0x00 0x00 0x00 0x00
          LE Encryption
          Extended Reject Indication
          Slave-initiated Features Exchange
          LE Ping
          LE Data Packet Length Extension
          LL Privacy
          Extended Scanner Filter Policies
          Channel Selection Algorithm #2

+++ net/nimble/controller/include/controller/ble_ll_conn.h
@@ -178,7 +178,7 @@ struct ble_ll_conn_sm
     uint8_t cur_ctrl_proc;
     uint8_t disconnect_reason;
     uint8_t rxd_disconnect_reason;
-    uint8_t common_features;        /* Just a uint8 for now */
+    uint32_t common_features;
     uint8_t vers_nr;
     uint16_t pending_ctrl_procs;
     uint16_t event_cntr;
